refactor: View를 Interface로 변경
