+++ src/main/kotlin/minesweeper/view/InputView.kt
@@ -1,27 +1,7 @@
 package minesweeper.view
 
-object InputView {
-    tailrec fun readHeight(): Int {
-        println("높이를 입력하세요.")
-        val userInput: String = readln()
-
-        if (userInput.toIntOrNull() != null) return userInput.toInt()
-        return readHeight()
-    }
-
-    tailrec fun readWidth(): Int {
-        println("너비를 입력하세요.")
-        val userInput: String = readln()
-
-        if (userInput.toIntOrNull() != null) return userInput.toInt()
-        return readWidth()
-    }
-
-    tailrec fun readMineCount(): Int {
-        println("지뢰는 몇개인가요?")
-        val userInput: String = readln()
-
-        if (userInput.toIntOrNull() != null) return userInput.toInt()
-        return readMineCount()
-    }
+interface InputView {
+    fun readHeight(): Int
+    fun readWidth(): Int
+    fun readMineCount(): Int
 }
